server/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-debug-toolbar==4.2.0
django-environ==0.11.2
psycopg2==2.9.9
setuptools==57.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
wheel==0.36.2
